+++ dashboard/src/main/webapp/src/pages/serverstatus/NodeListPage.vue
@@ -17,13 +17,18 @@
 
 <template>
   <div>
+    <el-button type="success" @click="dialogFormVisible = true" v-show="isExcludePage()"> Add Node </el-button>
+    <el-button type="danger" @click="handleDeleteNode" v-show="isExcludePage()">Delete({{ selectItemNum }})</el-button>
     <el-table
       :data="listPageData.tableData"
       height="800"
       style="width: 100%"
+      :row-key="rowKey"
       :default-sort="sortColumn"
       @sort-change="sortChangeEvent"
+      @selection-change="handlerSelectionChange"
     >
+      <el-table-column type="selection" width="55" v-show="isExcludePage()"/>
       <el-table-column prop="id" label="Id" min-width="140" sortable fixed />
       <el-table-column label="NodeInfo(ip:jetty/grpc/netty)" min-width="140" >
         <template v-slot="{ row }">
@@ -109,6 +114,30 @@
         </template>
       </el-table-column>
     </el-table>
+    <el-dialog
+        v-model="dialogFormVisible"
+        title="Please enter the server id list to be excluded:"
+        class="dialog-wrapper"
+        v-show="isExcludePage()"
+    >
+      <el-form>
+        <el-form-item :label-width="formLabelWidth">
+          <el-input
+              v-model="textarea"
+              class="textarea-wrapper"
+              :rows="10"
+              type="textarea"
+              placeholder="Please input"
+          />
+        </el-form-item>
+      </el-form>
+      <template #footer>
+        <div class="dialog-footer">
+          <el-button @click="dialogFormVisible = false">Cancel</el-button>
+          <el-button type="primary" @click="handleConfirmAddHandler"> Confirm </el-button>
+        </div>
+      </template>
+    </el-dialog>
   </div>
 </template>
 <script>
@@ -127,7 +156,10 @@ import {
   getShuffleServerConf,
   getShuffleServerMetrics,
   getShuffleServerPrometheusMetrics,
-  getShuffleServerStacks
+  getShuffleServerStacks,
+  addShuffleExcludeNodes,
+  getShuffleExcludeNodes,
+  removeShuffleExcludeNodes
 } from '@/api/api'
 
 export default {
@@ -137,6 +169,114 @@ export default {
     const sortColumn = reactive({})
     const listPageData = reactive({ tableData: [] })
     const isShowRemove = ref(false)
+    const dialogFormVisible = ref(false)
+    const formLabelWidth = '10%'
+    const textarea = ref('')
+    const selectItemNum = ref(0)
+    const rowKey = 'id'
+    const selectedRows = ref([])
+    function isExcludePage() {
+      return router.currentRoute.value.query.isExcludedPage === 'true';
+    }
+
+    async function getShuffleExcludeNodesPage() {
+      const res = await getShuffleExcludeNodes()
+      listPageData.tableData = res.data.data
+    }
+
+    // The system obtains data from global variables and requests the interface to obtain new data after data changes.
+    currentServerStore.$subscribe((mutable, state) => {
+      if (state.currentServer) {
+        getShuffleExcludeNodesPage()
+      }
+    })
+
+    onMounted(() => {
+      // If the coordinator address to request is not found in the global variable, the request is not initiated.
+      if (currentServerStore.currentServer) {
+        getShuffleExcludeNodesPage()
+      }
+    })
+
+    /**
+     * The following describes how to handle add events.
+     */
+    function handleConfirmAddHandler() {
+      dialogFormVisible.value = false
+      addShuffleExcludeNodesPage()
+      // Refreshing the number of blacklists.
+      updateTotalPage()
+      // Refreshing the Blacklist list.
+      getShuffleExcludeNodesPage()
+    }
+    async function addShuffleExcludeNodesPage() {
+      try {
+        const excludeNodes = textarea.value.split('\n').map((item) => item.trim())
+        const excludeNodesObj = { excludeNodes }
+        const res = await addShuffleExcludeNodes(excludeNodesObj)
+        if (res.status >= 200 && res.status < 300) {
+          if (res.data.data === 'success') {
+            ElMessage.success('Add successfully.')
+          } else {
+            ElMessage.error('Add failed.')
+          }
+        } else {
+          ElMessage.error('Failed to add due to server bad.')
+        }
+      } catch (err) {
+        ElMessage.error('Failed to add due to network exception.')
+      }
+    }
+    function handleDeleteNode() {
+      ElMessageBox.confirm('Are you sure about removing these nodes?', 'Warning', {
+        confirmButtonText: 'OK',
+        cancelButtonText: 'Cancel',
+        type: 'warning'
+      })
+          .then(() => {
+            if (selectedRows.value.length === 0) {
+              ElMessage({
+                type: 'info',
+                message: 'No node is selected, Nothing!'
+              })
+            } else {
+              const selectedIds = selectedRows.value.map((row) => row[rowKey])
+              deleteShuffleExcludedNodes(selectedIds)
+              // Refreshing the number of blacklists.
+              updateTotalPage()
+              // Refreshing the Blacklist list.
+              getShuffleExcludeNodesPage()
+              ElMessage({
+                type: 'success',
+                message: 'Delete completed'
+              })
+            }
+          })
+          .catch(() => {
+            ElMessage({
+              type: 'info',
+              message: 'Delete canceled'
+            })
+          })
+    }
+
+    async function deleteShuffleExcludedNodes(excludeNodes) {
+      try {
+        const excludeNodesObj = { excludeNodes }
+        const res = await removeShuffleExcludeNodes(excludeNodesObj)
+        if (res.status >= 200 && res.status < 300) {
+          if (res.data.data === 'success') {
+            ElMessage.success('Delete successfully.')
+          } else {
+            ElMessage.error('Delete failed.')
+          }
+        } else {
+          ElMessage.error('Failed to delete due to server bad.')
+        }
+      } catch (err) {
+        ElMessage.error('Failed to delete due to network exception.')
+      }
+    }
     async function deleteLostServer(row) {
       try {
         const params = { serverId: row.id }
@@ -174,6 +314,11 @@ export default {
 
     async function getShuffleUnhealthyListPage() {
       const res = await getShuffleUnhealthyList()
+      listPageData.tableData = res.data.data
+    }
+
+    async function getShuffleExcludeListPage() {
+      const res = await getShuffleExcludeNodes()
       listPageData.tableData = res.data.data
     }
 
@@ -279,6 +424,8 @@ export default {
       } else if (router.currentRoute.value.name === 'lostNodeList') {
         isShowRemove.value = true
         getShuffleLostListPage()
+      } else if (router.currentRoute.value.name === 'excludeNodeList') {
+        getShuffleExcludeListPage()
       }
     }
 
@@ -324,6 +471,10 @@ export default {
           // Cancelled
         })
     }
+    function handlerSelectionChange(selection) {
+      selectedRows.value = selection
+      selectItemNum.value = selectedRows.value.length
+    }
     return {
       listPageData,
       sortColumn,
@@ -335,8 +486,27 @@ export default {
       handlerServerMetrics,
       handlerServerStacks,
       memFormatter,
-      dateFormatter
+      dateFormatter,
+      dialogFormVisible,
+      formLabelWidth,
+      textarea,
+      selectItemNum,
+      handleDeleteNode,
+      getShuffleExcludeNodesPage,
+      handleConfirmAddHandler,
+      isExcludePage,
+      rowKey,
+      handlerSelectionChange
     }
   }
 }
 </script>
+
+<style>
+.textarea-wrapper {
+  width: 90%;
+}
+.dialog-wrapper {
+  width: 50%;
+}
+</style>
