+++ dashboard/src/main/webapp/src/router/index.js
@@ -20,7 +20,6 @@ import ApplicationPage from '@/pages/ApplicationPage.vue'
 import DashboardPage from '@/pages/DashboardPage.vue'
 import CoordinatorServerPage from '@/pages/CoordinatorServerPage.vue'
 import ShuffleServerPage from '@/pages/ShuffleServerPage.vue'
-import ExcludeNodeList from '@/pages/serverstatus/ExcludeNodeList'
 import NodeListPage from '@/pages/serverstatus/NodeListPage.vue'
 
 const routes = [
@@ -68,7 +67,7 @@ const routes = [
       {
         path: '/shuffleserverpage/excludeNodeList',
         name: 'excludeNodeList',
-        component: ExcludeNodeList
+        component: NodeListPage
       }
     ]
   },
